+++ Robotopia/Classes/StateComponent.h
@@ -25,7 +25,7 @@ public:
 	OVERRIDE bool init();
 	OVERRIDE void enter();
 	OVERRIDE void exit();
-	OVERRIDE void update(int dTime);
+	OVERRIDE void update(float dTime);
 
 	void	   addTransition(Transition addTransition);
 	void	   removeTransition(Transition remTranstion);
